Fix up the daemon startup so it accounts for persistent volumes, and have the swarm server run apache in non-debug mode.

+++ integration-test/src/docker/p4d/p4d-start.sh
@@ -6,12 +6,13 @@ P4D=/usr/local/sbin/p4d.bin
 
 if [ ! -d "$P4ROOT" ]; then
     mkdir -p "$P4ROOT" || exit 1
-    if [ -f /opt/sampledepot.tar.gz ]; then
-        ( cd /tmp && tar zxf /opt/sampledepot.tar.gz ) || exit 1
-        mv /tmp/PerforceSample/* "$P4ROOT" || exit 1
-        $P4D -jr "$P4ROOT"/checkpoint || exit 1
-        $P4D -xu || exit 1
-    fi
+fi
+if [ ! -d "$P4ROOT/HR" -a -f /opt/sampledepot.tar.gz ]; then
+    ( cd /tmp && tar zxf /opt/sampledepot.tar.gz ) || exit 1
+    rm -r "$P4ROOT"/* || exit 1
+    mv /tmp/PerforceSample/* "$P4ROOT" || exit 1
+    $P4D -jr "$P4ROOT"/checkpoint || exit 1
+    $P4D -xu || exit 1
 fi
 
 if [ "$P4PORT" = "ssl:1666" ]; then
